Prefer size_t over int where possible (#415)
This avoids needless truncation, which takes up multiple more instructions on some architectures, especially older ones.

+++ pppd/plugins/pppoatm/misc.c
@@ -20,7 +20,7 @@
 int __atmlib_fetch(const char **pos,...)
 {
     const char *value;
-    int ref_len,best_len,len;
+    size_t ref_len,best_len,len;
     int i,best;
     va_list ap;
